Code/requirements.txt
# ============================================================
# Blood Eye Project — Requirements
# Install with: pip install -r requirements.txt
# ============================================================

# Web Framework
Flask==3.1.3
Werkzeug==3.1.6

# Authentication
flask-dance==7.1.0
python-dotenv==1.1.1

# Database (built-in sqlite3 — no install needed)

# Image Processing
opencv-python==4.13.0.33
scikit-image==0.25.2
Pillow==11.3.0

# Machine Learning
scikit-learn==1.8.0
numpy==2.2.6
pandas==2.3.3
joblib==1.5.2

# Deep Learning
tensorflow==2.21.0

# PDF Report Generation
reportlab==4.4.10
qrcode==8.2

# Plotting (used in training and report)
matplotlib==3.10.7
seaborn==0.13.2
